Cast `m_Width` to float when dividing by `m_Height`

+++ Engine/Source/Runtime/Core/Window.h
@@ -65,7 +65,7 @@ namespace Nightly
 
 		NL_NODISCARD float GetAspectRatio() const
 		{
-			return m_Width / (float) m_Height;
+			return (float) m_Width / (float) m_Height;
 		}
 
 	private:
